refactor: Break Spot screen into separate widgets
