+++ dristi/lib/src/core/routes/app_router.dart
@@ -1,6 +1,5 @@
 import 'package:dristi/src/core/global_widgets/error_screen.dart';
 import 'package:dristi/src/core/global_widgets/web_view_screen.dart';
-import 'package:dristi/src/core/loggers/logger.dart';
 import 'package:dristi/src/core/routes/app_routes.dart';
 import 'package:dristi/src/core/routes/navigation_helper.dart';
 import 'package:dristi/src/core/utils/localization_ext.dart';
