+++ dristi/lib/src/features/destinations/presentation/pages/destinations_list_screen.dart
@@ -59,6 +59,7 @@ class _DestinationPageState
   }
 
   // For Future Usage
+  // ignore: unused_element
   Widget _buildAdvertisement() {
     return SliverToBoxAdapter(
       child: Padding(
