+++ dristi/lib/src/features/home/sliders/presentations/widgets/slider_builder.dart
@@ -134,7 +134,7 @@ class _ImageSliderBuilderState
                 border: Border.all(
                   color: currentSliderState == index
                       ? uiColors.primary
-                      : uiColors.primary.withOpacity(0.6),
+                      : uiColors.primary.withOpacity(0.2),
                 ),
               ),
             ),
