+++ dristi/lib/src/features/spot/presentation/widgets/spot_image.dart
@@ -1,0 +1,92 @@
+import 'package:dristi/src/core/base/base_stateful_widget.dart';
+import 'package:dristi/src/core/constants/app_assets.dart';
+import 'package:dristi/src/core/constants/app_values.dart';
+import 'package:dristi/src/core/utils/asset_image_view.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class SpotImage extends StatefulWidget {
+  const SpotImage({super.key});
+
+  @override
+  State<SpotImage> createState() => _SpotImageState();
+}
+
+class _SpotImageState extends BaseStatefulWidget<SpotImage> {
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      alignment: Alignment.topRight,
+      children: [
+        _buildBackgroundImage(),
+        _buildGalleryIcon(),
+        _buildDestinationText(),
+      ],
+    );
+  }
+
+  Widget _buildBackgroundImage() {
+    return ClipRRect(
+      child: AssetImageView(
+        fileName: Assets.coxsBazar,
+        fit: BoxFit.cover,
+        height: AppValues.dimen_500.h,
+        width: double.infinity,
+      ),
+    );
+  }
+
+  Widget _buildGalleryIcon() {
+    return Positioned(
+      bottom: AppValues.dimen_100.h,
+      left: AppValues.dimen_16.w,
+      child: GestureDetector(
+        onTap: () {},
+        child: Container(
+          padding: EdgeInsets.all(AppValues.dimen_3.r),
+          height: AppValues.dimen_75.r,
+          width: AppValues.dimen_75.r,
+          decoration: BoxDecoration(
+            color: uiColors.onImage,
+            borderRadius: BorderRadius.circular(AppValues.dimen_12.r),
+          ),
+          child: Stack(
+            children: [
+              ClipRRect(
+                borderRadius: BorderRadius.circular(AppValues.dimen_10.r),
+                child: SizedBox.fromSize(
+                  size: Size.fromRadius(AppValues.dimen_48.r),
+                  child: const AssetImageView(
+                    fileName: Assets.coxsBazar,
+                    fit: BoxFit.cover,
+                  ),
+                ),
+              ),
+              Center(
+                child: Icon(
+                  Icons.more_horiz_outlined,
+                  color: uiColors.onImage,
+                  size: AppValues.dimen_32.h,
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildDestinationText() {
+    return Positioned(
+      bottom: AppValues.dimen_60.h,
+      right: AppValues.dimen_16.w,
+      child: Transform.rotate(
+        angle: -15 * (3.1415926535 / 180),
+        child: Text(
+          "Saint Martins Island",
+          style: appTextStyles.onImageBoldShadow44,
+        ),
+      ),
+    );
+  }
+}
